important bug fix to new import code! (it was tripped up by very large numbers)
git-svn-id: http://192.168.1.10/repos/toto@2908 b53939d4-f08c-4181-96d4-b201aa9c61c4

+++ Toto/trunk/src/com/azquo/dataimport/ImportFileUtilities.java
@@ -203,7 +203,7 @@ class ImportFileUtilities {
                 if (dataFormat.contains("0") && dataFormat.replace("0","").isEmpty()){
                     returnString = df.formatCellValue(cell);
                 } else {
-                    returnString = returnNumber.intValue() + "";
+                    returnString = returnNumber.longValue() + "";
                 }
             }
             if (dataFormat.equals("h:mm") && returnString.length() == 4) {
